Extract TrelloBoard into trello_board.dart
